build(playwright): fix github hook
The main issue was that we were not using the playwright webserver
because we already run the tests with the dev server open.

So managed to filter when is it running with docker and when it's not,
to distinguish if it needs to run the dev server, and if it's the docker
url or localhost (outside docker).